Separate CSS from ToiletDetail
