+++ frontend/src/components/ToiletDetail/ToiletDetail.module.scss
@@ -2,45 +2,3 @@
   --bs-offcanvas-height: 80vh;
   min-height: 80vh;
 }
-
-.toilet-utilities {
-  display: flex;
-  justify-content: space-between;
-  max-width: 460px;
-  font-size: 0.9rem;
-}
-
-.box {
-  background-color: #f3f3f3;
-  border-style: solid;
-  border-radius: 0.375rem;
-  border-color: #f3f3f3;
-  margin-bottom: 28px;
-  padding: 12px;
-}
-
-.toilet-summary {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-bottom: 20px;
-  border: 2px solid;
-  border-radius: 0.375rem;
-  border-color: #f3f3f3;
-}
-
-.toilet-summary-item {
-  padding: 2px;
-  margin: 0;
-  text-align: center;
-  font-style: italic;
-}
-
-.toilet-placeholder-screen {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  height: 80%;
-}
