+++ frontend/src/components/ToiletDetail/ToiletDetailBody.module.scss
@@ -1,0 +1,32 @@
+.toilet-utilities {
+  display: flex;
+  justify-content: space-between;
+  max-width: 460px;
+  font-size: 0.9rem;
+}
+
+.box {
+  background-color: #f3f3f3;
+  border-style: solid;
+  border-radius: 0.375rem;
+  border-color: #f3f3f3;
+  margin-bottom: 28px;
+  padding: 12px;
+}
+
+.toilet-summary {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-bottom: 20px;
+  border: 2px solid;
+  border-radius: 0.375rem;
+  border-color: #f3f3f3;
+}
+
+.toilet-summary-item {
+  padding: 2px;
+  margin: 0;
+  text-align: center;
+  font-style: italic;
+}
