+++ frontend/src/components/ToiletDetail/ToiletPlaceholderScreen.module.scss
@@ -1,0 +1,8 @@
+.toilet-placeholder-screen {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+  height: 80%;
+}
